2025 CTF/FortID CTF 2025/MISC/Meta 2.0/requirements.txt
Flask==3.0.3
python-magic==0.4.27
